+++ lib/balancedtrees.dart
@@ -2,3 +2,4 @@ library balancedtrees;
 
 export 'package:balancedtrees/avltree/avltree.dart';
 export 'package:balancedtrees/bplustree/bplustree.dart';
+export 'package:balancedtrees/util/util.dart';